Add log param as artifact in mlflow utils.
